+++ inc/modules/pelion_stdlib.h
@@ -14,7 +14,7 @@ void
 pelion_sprintf(char *dest, const char *format, ...);
 
 void
-pelion_convert_numeric_to_float(long long num, unsigned short base,
+pelion_convert_numeric_to_float(long num, unsigned short base,
                                 char *dest, int max_dest_size);
 
 void
