+++ src/bare/modules/pelion_stdlib.c
@@ -473,7 +473,7 @@ pelion_sprintf(char *dest, const char *format, ...) {
  *  dest will be written with 1.234
  */
 void
-pelion_convert_numeric_to_float(long long num, unsigned short base,
+pelion_convert_numeric_to_float(long num, unsigned short base,
                                 char *dest, int max_dest_size) {
 
     char *prefix = "";
